:truck: Relocate password and users request types
